+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * st-form: added .small class to affect form inner inputs or input components individually.
 * st-dynamic-table: Allow to hide a field using the UI definitions
 * st-table: Allow to display "sticky" the hover menu
+* st-table: Close filter menu when user clicks outside
 
 **Fixed bugs:**